Move attributes into 'Attributes' folder
