+++ exmat/Attributes/ExStdLibBase.cs
@@ -41,6 +41,9 @@ namespace ExMat
         STATISTICS
     }
 
+    /// <summary>
+    /// Attribute to mark a class as a standard library of given <see cref="ExStdLibType"/> type
+    /// </summary>
     [AttributeUsage(AttributeTargets.Class, AllowMultiple = false, Inherited = false)]
     public class ExStdLibBase : Attribute
     {
