+++ exmat/Attributes/ExStdLibName.cs
@@ -2,6 +2,10 @@
 
 namespace ExMat
 {
+    /// <summary>
+    /// Attribute to set the name of a standard library
+    /// <para>This name is used by the users to refer to this library</para>
+    /// </summary>
     [AttributeUsage(AttributeTargets.Class, AllowMultiple = false, Inherited = false)]
     public class ExStdLibName : Attribute
     {
